backend/requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
bcrypt==4.3.0
PyMySQL
SQLAlchemy
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
PyJWT
argon2_cffi
email-validator
pytz
pyodbc